backend/todos/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter

from .views import TodoViewSet

router = DefaultRouter(trailing_slash=False)
router.register(r'', TodoViewSet, basename='todo')

urlpatterns = [
    path('', include(router.urls)),
]
