backend/requirements.txt
Django>=4.2,<5
gunicorn>=21.0
psycopg[binary]>=3.1
djangorestframework>=3.14
